Reorganize calling center
- Isolate parts in different files
- Create logical group of methods in calling center
- Add documentation
