+++ Source/Calling/CallClosedReason.swift
@@ -19,13 +19,20 @@
 import Foundation
 import avs
 
+/**
+ * Reasons why a call can be terminated.
+ */
+
 public enum CallClosedReason : Int32 {
+
     /// Ongoing call was closed by remote or self user
     case normal
     /// Incoming call was canceled by remote
     case canceled
     /// Incoming call was answered on another device
     case anweredElsewhere
+    /// Incoming call was rejected on another device
+    case rejectedElsewhere
     /// Outgoing call timed out
     case timeout
     /// Ongoing call lost media and was closed
@@ -41,6 +48,14 @@ public enum CallClosedReason : Int32 {
     /// Call was closed for an unknown reason. This is most likely a bug.
     case unknown
 
+    // MARK: - Briding
+
+    /**
+     * Creates the call closed reason from the AVS flag.
+     * - parameter wcall_reason: The flag
+     * - returns: The decoded reason, or `.unknown` if the flag couldn't be processed.
+     */
+
     init(wcall_reason: Int32) {
         switch wcall_reason {
         case WCALL_REASON_NORMAL:
@@ -49,6 +64,8 @@ public enum CallClosedReason : Int32 {
             self = .canceled
         case WCALL_REASON_ANSWERED_ELSEWHERE:
             self = .anweredElsewhere
+        case WCALL_REASON_REJECTED:
+            self = .rejectedElsewhere
         case WCALL_REASON_TIMEOUT:
             self = .timeout
         case WCALL_REASON_LOST_MEDIA:
@@ -64,6 +81,7 @@ public enum CallClosedReason : Int32 {
         }
     }
 
+    /// The raw flag for the call end.
     var wcall_reason : Int32 {
         switch self {
         case .normal:
@@ -72,6 +90,8 @@ public enum CallClosedReason : Int32 {
             return WCALL_REASON_CANCELED
         case .anweredElsewhere:
             return WCALL_REASON_ANSWERED_ELSEWHERE
+        case .rejectedElsewhere:
+            return WCALL_REASON_REJECTED
         case .timeout:
             return WCALL_REASON_TIMEOUT
         case .lostMedia:
