+++ Source/Calling/FlowManager.swift
@@ -46,43 +46,45 @@ public class FlowManager : NSObject, FlowManagerType {
     
 }
 
+// MARK: - AVSFlowManagerDelegate
+
 extension FlowManager : AVSFlowManagerDelegate {
     
-    
     public static func logMessage(_ msg: String!) {
-        
+        // no-op
     }
     
     public func request(withPath path: String!, method: String!, mediaType mtype: String!, content: Data!, context ctx: UnsafeRawPointer!) -> Bool {
+        // no-op
         return false
     }
     
     public func didEstablishMedia(inConversation convid: String!) {
-        
+        // no-op
     }
     
     public func didEstablishMedia(inConversation convid: String!, forUser userid: String!) {
-        
+        // no-op
     }
     
     public func setFlowManagerActivityState(_ activityState: AVSFlowActivityState) {
-        
+        // no-op
     }
     
     public func networkQuality(_ q: Float, conversation convid: String!) {
-        
+        // no-op
     }
     
     public func mediaWarning(onConversation convId: String!) {
-        
+        // no-op
     }
     
     public func errorHandler(_ err: Int32, conversationId convid: String!, context ctx: UnsafeRawPointer!) {
-        
+        // no-op
     }
     
     public func didUpdateVolume(_ volume: Double, conversationId convid: String!, participantId: String!) {
-
+        // no-op
     }
     
 }
