+++ Source/Calling/SystemMessageCallObserver.swift
@@ -16,13 +16,11 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-
 import Foundation
-
 
 private let log = ZMSLog(tag: "Calling System Message")
 
-/// Inserts a calling system message for V3 calls
+/// Inserts a calling system message for V3 calls.
 final class CallSystemMessageGenerator: NSObject {
     
     var startDateByConversation = [ZMConversation: Date]()
